Rewrote Hero.enumerateEntities().
Increased readability and (as a bonus) reduced time complexity.

Was O(mn) - where m is the number of different names and n is the number of Entities.
O(m + n) now. This way may even be slower as m and n are usually pretty small.
Performance wasn't an issue (neither was time complexity, but w/e).