Fixes: picked up item had inventory set to null.
This happened because the Item was first added to the Creature's inventory
(setting the inventory field to that Creature's inventory) and then removed from
the Location's inventory, setting the inventory field to null.

Creature inventory size and weight limit are now loaded from resource files.

If 'debug give' cannot add the Item to the player's inventory, it drops it in
the current location.

Several parts of the code were clarified.

Adding items to and dropping items from the Hero's inventory now use the
qualified name of the item.

"Natural" items such as Stick and Stone are no longer repairable.