Hero gets a Potato at start (fixed a bug).
Fixes: Hero can't start with items sometimes.

This happened because the Date required to create those items was taken
from the GameState in Game. When it was not null, it worked (but produced
wrong results), when it wasn't, you got a NullPointerException.

Now the game sets the GameState in Game to null when creating a new
GameState from the 'new' command, this way bugs of this kind will be easy
to find.

Made some refactorings in CreatureFactory to be able to create the Hero
with Items. Gave Seth a Potato as a gift.