Improved attack algorithms organization.
Eliminated a good amount of duplicated code.
AttackAlgorithms no longer uses a static initializer block.
Lazy initialization is being used even there now.
This also avoids unwanted ExceptionInInitializerError (not descriptive enough to me).
Items not use Percentage to represent hit rate, this is better as it improves readability in JSON.
Fixed a bug that allowed the player to commit suicide.

Closes #117