Fixed the repeated dungeon entrance bug
This is a major fix and should be preserved.

Essentially what happened was that the set of points that kept track
of the existing entrances was a static field. This means that when a
new World was generated, the DungeonDistributor of its WorldGenerator
would already have some entrances registered. Just making the field
non-static fixed the problem. Each DungeonDistributor (every
WorldGenerator has one of these) should have its own entrances
Set<Point>.

The main consequence of this bug was that subsequent worlds would not
have as many dungeons because the isIsolatedEnough method would see
dungeon entrances of the previous worlds and so the number of points
that would be allowed to have a dungeon was much smaller.

Fixes #160